+++ SolutionItems/Properties/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("Wealth Economy - Solution Items")]
-[assembly: AssemblyDescription("General files for the solution")]